Commit remaining session changes
