+++ packages/web/src/views/SessionDetailView.vue
@@ -128,6 +128,7 @@ import { useUiStore } from '../stores/ui.js';
 import { useKanbanStore } from '../stores/kanban.js';
 import { useSessionPolling } from '../composables/useSessionPolling.js';
 import { useSessionInitializer } from '../composables/useSessionInitializer.js';
+import { useSessionTree } from '../composables/useSessionTree.js';
 import ChangesTab from '../components/ChangesTab.vue';
 import CanvasTab from '../components/CanvasTab.vue';
 import SummaryTab from '../components/SummaryTab.vue';
@@ -139,10 +140,7 @@ import SessionChatOverlay from '../components/SessionChatOverlay.vue';
 import SessionChatContent from '../components/SessionChatContent.vue';
 import ArchiveConfirmModal from '../components/ArchiveConfirmModal.vue';
 import { useCommandButtonsStore } from '../stores/commandButtons.js';
-import { api } from '../composables/useApi.js';
 import { useWebSocket } from '../composables/useWebSocket.js';
-import { sortSessionChain } from '../utils/sessionPickerRecency.js';
-import { isSessionDetailDesktop } from '../composables/useSessionDetailBreakpoint.js';
 import { WS_MESSAGE_TYPES } from '@circuschief/shared';
 
 const route = useRoute();
@@ -157,18 +155,11 @@ const kanbanStore = useKanbanStore();
 
 const { send, on, off } = useWebSocket();
 
-// Track current project subscription for cleanup on route change and unmount
 let currentProjectSubscriptionId = null;
 
-// Reactive session ID that tracks route changes
-// Used by the polling composable to track the current session
 const currentSessionId = ref(route.params.id);
-
-// Set viewedSessionId immediately so that stale in-flight fetchSession requests
-// from a previous session's polling cannot overwrite currentSession.
 sessionsStore.viewedSessionId = route.params.id;
 
-// Use composable for polling and file changes
 const {
   hasChanges,
   changesFileCount,
@@ -184,24 +175,10 @@ const {
 
 const summary = ref(null);
 const isDeleting = ref(false);
-const chatOverlayOpen = ref(false);
 const showArchiveModal = ref(false);
 const archiving = ref(false);
 
-// Session ID to pass to the overlay - resolves to running child if present
-const overlaySessionId = ref(route.params.id);
-const preferredOverlaySessionId = ref(null);
-const preferredOverlaySession = ref(null);
-
-// Session chain state (lifted from SessionChatOverlay)
-const sessionChain = ref([]);
-const summariesMap = ref({});
-const hasDescendants = computed(() => sessionChain.value.length > 1);
-
-// Readiness signal for E2E tests. Flips to true once the store has loaded
-// the current session AND the session-tree initial fetch has resolved.
-// Resets to false when the route id changes, so tests can wait on it after
-// navigation without racing the hydration lifecycle.
+// Readiness signal for E2E tests
 const sessionChainReady = ref(false);
 const isReady = computed(() =>
   !sessionsStore.loading &&
@@ -209,269 +186,26 @@ const isReady = computed(() =>
   sessionChainReady.value,
 );
 
-/**
- * Merge project sessions into the store without triggering loading state.
- */
-async function mergeProjectSessionsToStore(projectId) {
-  try {
-    const projectSessions = await api.getProjectSessions(projectId, false, null);
-    for (const s of projectSessions) {
-      const idx = sessionsStore.sessions.findIndex(existing => existing.id === s.id);
-      if (idx >= 0) {
-        sessionsStore.sessions[idx] = s;
-      } else {
-        sessionsStore.sessions.push(s);
-      }
-    }
-  } catch {
-    // Not critical if project sessions fail to load
-  }
-}
+// Session tree composable (chain building, overlay target, chat navigation)
+const {
+  chatOverlayOpen,
+  overlaySessionId,
+  sessionChain,
+  summariesMap,
+  isSessionActive,
+  activeSessionStatus,
+  buildSessionChain,
+  resolveOverlayTarget,
+  openChatDestination,
+  handleOverlayOpen,
+  handleScheduledSessionClick,
+  handleSessionUpdated,
+  handleSessionCreated,
+  handleOverlaySessionCreated,
+  handleOverlayClose,
+  resetPreferred,
+} = useSessionTree(currentSessionId, sessionChainReady);
 
-/**
- * Find the root session, handling fallback cases.
- * @returns {{ root: object|null, earlyReturn: object[]|null }}
- */
-function findRootSession(sessionId) {
-  const root = sessionsStore.getRootSession(sessionId);
-  if (root) return { root, earlyReturn: null };
-
-  const current = sessionsStore.getSessionById(sessionId) || sessionsStore.currentSession;
-  if (current && !current.parentSessionId) return { root: current, earlyReturn: null };
-  if (current) return { root: null, earlyReturn: [{ session: current, depth: 0 }] };
-  return { root: null, earlyReturn: null };
-}
-
-/**
- * Walk the tree depth-first from root, collecting {session, depth} entries.
- */
-function collectTreeDepthFirst(root) {
-  const tree = [];
-  function walk(session, depth) {
-    tree.push({ session, depth });
-    const children = sessionsStore.getChildSessions(session.id);
-    for (const child of children) walk(child, depth + 1);
-  }
-  walk(root, 0);
-  return tree;
-}
-
-/**
- * Build the full session tree from root, flattened depth-first with depth info.
- * Fetches project sessions and summaries for each session in the tree.
- */
-async function buildSessionChain() {
-  const sessionId = currentSessionId.value;
-  const existingSession = sessionsStore.getSessionById(sessionId) || sessionsStore.currentSession;
-  if (!existingSession) {
-    try { await sessionsStore.fetchSession(sessionId, false); } catch { return; }
-  }
-
-  const session = sessionsStore.getSessionById(sessionId) || sessionsStore.currentSession;
-  if (session?.projectId) await mergeProjectSessionsToStore(session.projectId);
-
-  const { root, earlyReturn } = findRootSession(sessionId);
-  if (earlyReturn) { sessionChain.value = sortSessionChain(earlyReturn); return; }
-  if (!root) return;
-
-  const tree = sortSessionChain(collectTreeDepthFirst(root));
-  sessionChain.value = tree;
-
-  // Fetch summaries for all sessions in the tree (non-blocking)
-  for (const entry of tree) {
-    if (!summariesMap.value[entry.session.id]) {
-      api.getSessionSummary(entry.session.id)
-        .then(fetchedSummary => { if (fetchedSummary) summariesMap.value = { ...summariesMap.value, [entry.session.id]: fetchedSummary }; })
-        .catch(() => { /* Summaries are not critical */ });
-    }
-  }
-}
-
-/**
- * Resolve the overlay target session ID.
- * Priority order:
- * 1. Running/starting children (most recent picker recency)
- * 2. Session with the most recent picker recency
- * 3. Current session (fallback)
- */
-function resolveOverlayTarget() {
-  const chain = sessionChain.value;
-
-  if (preferredOverlaySessionId.value) {
-    const preferred = chain.find(entry => entry.session.id === preferredOverlaySessionId.value);
-    if (preferred) {
-      overlaySessionId.value = preferred.session.id;
-      return;
-    }
-    if (preferredOverlaySession.value?.id === preferredOverlaySessionId.value) {
-      overlaySessionId.value = preferredOverlaySession.value.id;
-      return;
-    }
-  }
-
-  // No children — use the current session
-  if (chain.length <= 1) {
-    overlaySessionId.value = currentSessionId.value;
-    return;
-  }
-
-  // 1) Prefer running/starting children (skip the root at index 0)
-  const runningChildren = chain
-    .filter(entry => entry.session.status === 'running' || entry.session.status === 'starting')
-    .filter(entry => entry.session.id !== currentSessionId.value);
-
-  if (runningChildren.length > 0) {
-    const sorted = sortSessionChain(runningChildren);
-    overlaySessionId.value = sorted[0].session.id;
-    return;
-  }
-
-  const withActivity = sortSessionChain(chain)
-    .filter(entry => entry.pickerTimestamp);
-
-  if (withActivity.length > 0) {
-    overlaySessionId.value = withActivity[0].session.id;
-    return;
-  }
-
-  // 3) No conversation activity anywhere — use the current session
-  overlaySessionId.value = currentSessionId.value;
-}
-
-/**
- * Handle SESSION_UPDATED WebSocket events.
- * When a session in our tree changes status, update the sessionChain snapshot
- * so that isSessionActive and activeSessionStatus recompute correctly.
- */
-function handleSessionUpdated(msg) {
-  const updatedSession = msg.session;
-  if (!updatedSession) return;
-
-  // Update the session in sessionChain if it's part of our tree
-  const idx = sessionChain.value.findIndex(
-    entry => entry.session.id === updatedSession.id
-  );
-  if (idx >= 0) {
-    const updatedEntries = [...sessionChain.value];
-    // Replace the stale snapshot with the updated one, preserving depth.
-    updatedEntries[idx] = {
-      ...sessionChain.value[idx],
-      session: updatedSession,
-    };
-    sessionChain.value = sortSessionChain(updatedEntries);
-  }
-}
-
-/**
- * Handle SESSION_CREATED WebSocket events.
- * When the overlay is closed and a new child session is created in our tree,
- * update overlaySessionId so the next open navigates to the new child.
- */
-function handleSessionCreated(msg) {
-  const projectId = sessionsStore.currentSession?.projectId;
-  if (!projectId || msg.projectId !== projectId) return;
-
-  const newSession = msg.session;
-  if (!newSession?.parentSessionId) return;
-
-  // Only act when overlay is closed
-  if (chatOverlayOpen.value) return;
-
-  // Check if the new session's parent is in our session tree
-  const isChildOfTree = sessionChain.value.some(
-    entry => entry.session.id === newSession.parentSessionId
-  );
-  if (!isChildOfTree) return;
-
-  // Add to store so getters work
-  sessionsStore.addSessionToList(newSession);
-
-  // Update overlay target BEFORE the async chain rebuild so it's set immediately
-  if (newSession.status === 'running' || newSession.status === 'starting') {
-    overlaySessionId.value = newSession.id;
-  }
-
-  // Rebuild the tree to include the new child (async, fire-and-forget)
-  buildSessionChain();
-}
-
-async function openChatDestination({ targetSessionId = null, replaceQuery = false } = {}) {
-  if (targetSessionId) {
-    preferredOverlaySessionId.value = targetSessionId;
-    overlaySessionId.value = targetSessionId;
-  } else {
-    resolveOverlayTarget();
-  }
-
-  if (isSessionDetailDesktop()) {
-    chatOverlayOpen.value = false;
-    const path = `/sessions/${currentSessionId.value}/chat`;
-    const navigation = { path, query: {} };
-    if (replaceQuery) {
-      await router.replace(navigation);
-    } else if (route.path !== path) {
-      await router.push(path);
-    }
-    return;
-  }
-
-  chatOverlayOpen.value = true;
-  if (replaceQuery) {
-    await router.replace({ path: route.path, query: {} });
-  }
-}
-
-function handleOverlayOpen() {
-  openChatDestination();
-}
-
-async function handleScheduledSessionClick(sessionId) {
-  await buildSessionChain();
-  await openChatDestination({ targetSessionId: sessionId });
-}
-
-async function handleOverlaySessionCreated(session) {
-  const createdSession = typeof session === 'string'
-    ? sessionsStore.getSessionById(session)
-    : session;
-  const sessionId = typeof session === 'string' ? session : session?.id;
-  if (!sessionId) return;
-
-  preferredOverlaySessionId.value = sessionId;
-  preferredOverlaySession.value = createdSession || null;
-  overlaySessionId.value = sessionId;
-
-  if (createdSession) {
-    sessionsStore.addSessionToList(createdSession);
-  }
-
-  await buildSessionChain();
-  if (
-    preferredOverlaySession.value &&
-    !sessionChain.value.some(entry => entry.session.id === sessionId)
-  ) {
-    const parentEntry = sessionChain.value.find(
-      entry => entry.session.id === preferredOverlaySession.value.parentSessionId
-    );
-    sessionChain.value = sortSessionChain([
-      { session: preferredOverlaySession.value, depth: parentEntry ? parentEntry.depth + 1 : 0 },
-      ...sessionChain.value,
-    ]);
-  }
-  resolveOverlayTarget();
-}
-
-async function handleOverlayClose() {
-  chatOverlayOpen.value = false;
-  sessionsStore.viewedSessionId = currentSessionId.value;
-
-  // With overlay store isolation the main store is never touched by the overlay.
-  // A lightweight re-fetch picks up any server-side changes made while the overlay
-  // was open (e.g. status changes, new messages from other clients).
-  await sessionsStore.fetchSession(currentSessionId.value, false);
-}
-
-// Use composable for session initialization and WebSocket management
 const { cleanup, initializeSession } = useSessionInitializer({
   summary,
   hasChanges,
@@ -495,9 +229,7 @@ watch(
   { immediate: true }
 );
 
-// Command button status indicators for real-time updates (mirrors SessionCard behavior)
 const buttonStatusesToDisplay = computed(() => {
-  // Access commandRunVersion to establish Vue dependency tracking.
   // eslint-disable-next-line no-unused-vars
   const _version = sessionsStore.commandRunVersion;
 
@@ -519,39 +251,15 @@ const buttonStatusesToDisplay = computed(() => {
     }));
 });
 
-const isSessionActive = computed(() => {
-  // Check current session first (fast path)
-  const status = sessionsStore.currentSession?.status;
-  if (status === 'running' || status === 'starting') return true;
-
-  // Also check if any session in the chain (descendants) is running/starting
-  return sessionChain.value.some(entry =>
-    entry.session.status === 'running' || entry.session.status === 'starting'
-  );
-});
-
-const activeSessionStatus = computed(() => {
-  const currentStatus = sessionsStore.currentSession?.status;
-  if (currentStatus === 'running' || currentStatus === 'starting') return currentStatus;
-
-  // Find the first running/starting session in the chain
-  const active = sessionChain.value.find(entry =>
-    entry.session.status === 'running' || entry.session.status === 'starting'
-  );
-  return active?.session.status || currentStatus;
-});
-
 const tabs = computed(() => [
   { id: 'summary', label: 'Summary' },
+  { id: 'chat', label: 'Chat', desktopOnly: true },
   { id: 'changes', label: changesFileCount.value > 0 ? `Changes (${changesFileCount.value})` : 'Changes' },
   { id: 'canvas', label: canvasStore.groupedItems.length > 0 ? `Canvas (${canvasStore.groupedItems.length})` : 'Canvas' },
   { id: 'commands', label: 'Commands' },
-  { id: 'circus-time', label: 'Circus Time' },
-  { id: 'chat', label: 'Chat', desktopOnly: true }
+  { id: 'circus-time', label: 'Circus Time' }
 ]);
 
-// Watch for status changes from any source (optimistic updates, WebSocket, etc.)
-// This ensures polling starts even when status is updated directly in the store
 watch(
   () => sessionsStore.currentSession?.status,
   (newStatus, oldStatus) => {
@@ -564,36 +272,29 @@ watch(
 );
 
 onMounted(async () => {
-  // Register the SESSION_CREATED and SESSION_UPDATED handlers
   on(WS_MESSAGE_TYPES.SESSION_CREATED, handleSessionCreated);
   on(WS_MESSAGE_TYPES.SESSION_UPDATED, handleSessionUpdated);
 
-  // Initialize the session with WebSocket subscription and data fetching
   await initializeSession(currentSessionId.value);
 
-  // Fetch project data so ArchiveConfirmModal can check for cleanup scripts
   const projectId = sessionsStore.currentSession?.projectId;
   if (projectId && !projectsStore.currentProject) {
     await projectsStore.fetchProject(projectId);
   }
 
-  // Build session chain BEFORE resolving overlay target (resolveOverlayTarget reads sessionChain)
   await buildSessionChain();
   sessionChainReady.value = true;
   resolveOverlayTarget();
 
-  // Subscribe to project channel for SESSION_CREATED events
   if (projectId) {
     send(WS_MESSAGE_TYPES.SUBSCRIBE_PROJECT, { projectId });
     currentProjectSubscriptionId = projectId;
   }
 
-  // Auto-open tree overlay if requested via query param (e.g., after new session creation)
   if (route.query.overlay === 'open') {
     await openChatDestination({ replaceQuery: true });
   }
 
-  // Fetch kanban board so SessionHeaderPanel can show lane chip
   const session = sessionsStore.currentSession;
   if (session?.projectId) {
     kanbanStore.fetchBoard(session.projectId).catch(err => {
@@ -602,31 +303,22 @@ onMounted(async () => {
   }
 });
 
-// Watch for session changes within the same component (e.g., navigating between sessions)
-// CRITICAL: This fixes the WebSocket subscription leak bug
 watch(
   () => route.params.id,
   async (newSessionId, oldSessionId) => {
     if (newSessionId && newSessionId !== oldSessionId) {
-      // Set viewedSessionId BEFORE cleanup so that any in-flight fetchSession
-      // from the old session's polling is discarded.
       sessionsStore.viewedSessionId = newSessionId;
-      // Reset readiness so tests (and any caller watching isReady) observe
-      // the hydration transition for the new session.
       sessionChainReady.value = false;
-      preferredOverlaySessionId.value = null;
-      preferredOverlaySession.value = null;
+      resetPreferred();
       cleanup();
       currentSessionId.value = newSessionId;
       await initializeSession(newSessionId);
 
-      // Fetch project data if project changed so ArchiveConfirmModal can check for cleanup scripts
       const newProjectId = sessionsStore.currentSession?.projectId;
       if (newProjectId && projectsStore.currentProject?.id !== newProjectId) {
         await projectsStore.fetchProject(newProjectId);
       }
 
-      // Build session chain BEFORE resolving overlay target (resolveOverlayTarget reads sessionChain)
       await buildSessionChain();
       sessionChainReady.value = true;
       resolveOverlayTarget();
@@ -634,7 +326,6 @@ watch(
         await openChatDestination({ replaceQuery: true });
       }
 
-      // Update project subscription if project changed
       if (newProjectId !== currentProjectSubscriptionId) {
         if (currentProjectSubscriptionId) {
           send(WS_MESSAGE_TYPES.UNSUBSCRIBE_PROJECT, { projectId: currentProjectSubscriptionId });
@@ -656,7 +347,6 @@ watch(
   }
 );
 
-// Watch for conversation changes to refetch todos (scoped to conversation)
 watch(
   () => sessionsStore.activeConversationId,
   async (newConvId, oldConvId) => {
@@ -667,7 +357,6 @@ watch(
   }
 );
 
-// Handle component reactivation (keep-alive) - refresh data when view becomes active again
 onActivated(() => {
   if (currentSessionId.value) {
     sessionsStore.fetchConversations(currentSessionId.value);
@@ -676,16 +365,12 @@ onActivated(() => {
 
 onUnmounted(() => {
   cleanup();
-  // Unsubscribe from project channel
   if (currentProjectSubscriptionId) {
     send(WS_MESSAGE_TYPES.UNSUBSCRIBE_PROJECT, { projectId: currentProjectSubscriptionId });
     currentProjectSubscriptionId = null;
   }
-  // Remove the SESSION_CREATED and SESSION_UPDATED handlers
   off(WS_MESSAGE_TYPES.SESSION_CREATED, handleSessionCreated);
   off(WS_MESSAGE_TYPES.SESSION_UPDATED, handleSessionUpdated);
-  // Clear viewedSessionId so other views (e.g., SessionListView) can use
-  // fetchSession without the guard blocking them.
   sessionsStore.viewedSessionId = null;
 });
 
@@ -736,7 +421,6 @@ async function handleArchive() {
   const isArchived = sessionsStore.currentSession?.archived;
 
   if (isArchived) {
-    // Unarchive path: keep simple confirm dialog
     if (!confirm('Restore this session to active?')) {
       return;
     }
@@ -753,7 +437,6 @@ async function handleArchive() {
       uiStore.error(err.message);
     }
   } else {
-    // Archive path: show modal with cleanup option
     showArchiveModal.value = true;
   }
 }
@@ -812,7 +495,6 @@ async function handleCopySessionId() {
   }
 }
 
-// Expose for testing
 defineExpose({
   overlaySessionId,
   chatOverlayOpen,
