+++ tests/e2e/scheduling-reschedule.spec.ts
@@ -567,9 +567,9 @@ test.describe('Category 5: Scheduling UI Components', () => {
     await navigateAndWait(page, `/sessions/${session.id}/summary`);
     await openSessionOverlay(page);
 
-    // Click "Edit Schedule" button (scoped to overlay to avoid duplicate from SummaryTab)
-    const overlay = page.getByTestId('session-chat-overlay');
-    const editButton = overlay.getByRole('button', { name: 'Edit Schedule' });
+    // Click "Edit Schedule" button (scoped to chat content to avoid duplicate from SummaryTab)
+    const chatContent = page.locator('.session-chat-content');
+    const editButton = chatContent.getByRole('button', { name: 'Edit Schedule' });
     await expect(editButton).toBeVisible({ timeout: 10000 });
     await editButton.click();
