+++ tests/e2e/session-chat-overlay-scroll.spec.ts
@@ -32,19 +32,32 @@ test.describe('Session Chat Overlay Scroll Behavior', () => {
     await cleanupCreatedResources();
   });
 
-  // Helper to navigate and open the overlay
+  // Helper to navigate and open the overlay. Wider viewport cases open at a
+  // mobile width first because the product intentionally routes >=641px to the
+  // embedded Chat tab, then restore the requested viewport before assertions.
   async function openOverlay(page: any, sessionId: string) {
+    const originalViewport = page.viewportSize();
     await navigateAndWait(page, `/sessions/${sessionId}`, {
       waitFor: '.session-detail',
       timeout: 15000,
     });
     const handle = page.locator('[data-testid="session-chat-handle"]');
-    await expect(handle).toBeVisible({ timeout: 10000 });
+    await handle.waitFor({ state: 'attached', timeout: 10000 });
+
+    if (originalViewport && originalViewport.width >= 641) {
+      await page.setViewportSize({ width: 390, height: originalViewport.height });
+    }
     await handle.click();
+
     const overlay = page.locator('[data-testid="session-chat-overlay"]');
-    await expect(overlay).toBeVisible({ timeout: 5000 });
-    // Wait for slide-in animation to complete
-    await page.waitForTimeout(400);
+    await overlay.waitFor({ state: 'visible', timeout: 5000 });
+    if (originalViewport && originalViewport.width >= 641) {
+      await page.setViewportSize(originalViewport);
+      await page.evaluate(() => window.dispatchEvent(new Event('resize')));
+    }
+    // Wait for overlay header to confirm full render, then allow animation to settle.
+    await overlay.locator('.overlay-header').waitFor({ state: 'visible', timeout: 5000 });
+    await page.waitForTimeout(300);
     return overlay;
   }
 
@@ -310,10 +323,13 @@ test.describe('Session Chat Overlay Scroll Behavior', () => {
     if (!btnBox) throw new Error('btnBox is null');
 
     // Same assertions as the no-cost case: 44×44, in viewport, in .overlay-body, no .input-form overlap, topmost.
+    const vp = page.viewportSize();
+    const vpWidth = vp?.width ?? 375;
+    const vpHeight = vp?.height ?? 667;
     expect(btnBox.width).toBeGreaterThanOrEqual(44);
     expect(btnBox.height).toBeGreaterThanOrEqual(44);
-    expect(btnBox.x + btnBox.width).toBeLessThanOrEqual(375);
-    expect(btnBox.y + btnBox.height).toBeLessThanOrEqual(667);
+    expect(btnBox.x + btnBox.width).toBeLessThanOrEqual(vpWidth);
+    expect(btnBox.y + btnBox.height).toBeLessThanOrEqual(vpHeight);
     const bodyBox = await overlay.locator('.overlay-body').boundingBox();
     expect(bodyBox).not.toBeNull();
     if (!bodyBox) throw new Error('bodyBox is null');
